feat: add L1Tabs component and reorganize sidebar navigation structure
