+++ packages/plugins/plugin-navtree/src/components/Sidebar/index.ts
@@ -4,3 +4,4 @@
 
 export * from './L0Menu';
 export * from './L1Panel';
+export * from './L1Tabs';
